Now provide a list of IP library locations that can be searched
Fixed issue with the linking of the external BDC interface and the internal interfaces

Moving the XSA parsing class into the utils module

Passed through black to fix line-length issues

Added the licenses

Using the smaller license format

Removing author, copyright, and email tags

Reverting the pass through black to fix line spacing issues

Better names for the Bdc Metadata classes

Fixing import issue when running on build machine

Fixed issue caused by BDC metadata class renaming
